moved js out of html to separate script
